+++ tests/test_midi.py
@@ -9,9 +9,9 @@ class TestMIDIFile(unittest.TestCase):
 
     def test_init(self):
         m = MIDIFile()
-        self.assertEquals(m.step_counter, 0)
+        self.assertEqual(m.filename, 'example.mid')
         m = MIDIFile(filename="test.mid")
-        self.assertEquals(m.filename, 'test.mid')
+        self.assertEqual(m.filename, 'test.mid')
 
     def test_create(self):
         notes = [1, 2, 3, 4, 5, 6, 7, 8, 9, 10, 11]